+++ src/Bitmaps/Icons.h
@@ -1,0 +1,38 @@
+// KMMX Accent Color RGB888
+static const uint8_t accColor[] PROGMEM = {
+    0xff, 0x99, 0x66, 0xff, 0x97, 0x66, 0xff, 0x96, 0x65, 0xff, 0x94, 0x64,
+    0xff, 0x91, 0x64, 0xff, 0x8f, 0x63, 0xff, 0x8c, 0x62, 0xff, 0x89, 0x62,
+    0xff, 0x86, 0x61, 0xff, 0x83, 0x60, 0xff, 0x80, 0x5f, 0xff, 0x7c, 0x5e,
+    0xff, 0x79, 0x5d, 0xff, 0x75, 0x5c, 0xfe, 0x71, 0x5b, 0xfe, 0x6d, 0x5a,
+    0xfe, 0x69, 0x59, 0xfe, 0x65, 0x58, 0xfd, 0x60, 0x57, 0xfd, 0x5c, 0x56,
+    0xfd, 0x57, 0x55, 0xfd, 0x53, 0x54, 0xfc, 0x4e, 0x53, 0xfc, 0x4a, 0x52,
+    0xfc, 0x45, 0x51, 0xfb, 0x40, 0x51, 0xfb, 0x3b, 0x50, 0xfa, 0x36, 0x4f,
+    0xfa, 0x31, 0x4e, 0xfa, 0x2c, 0x4e, 0xf9, 0x27, 0x4d, 0xf9, 0x22, 0x4d };
+
+//  ********************* Facial icons in 8bit Greyscale *********************
+// ----------- Nose 10x6 -----------
+static const uint8_t PROGMEM noseDefault[] = {
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x0d, 0x6f, 0xd0, 0xcc, 0x59, 0x06, 0x00, 0x00, 0x00, 0x00, 0x36, 0xa9,
+    0xbc, 0xf0, 0xff, 0x6b, 0x00, 0x00, 0x00, 0x00, 0x16, 0x11, 0x09, 0x2f,
+    0xec, 0xe5, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x4b, 0x7e,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 };
+
+// 8x6
+static const uint8_t PROGMEM heart[] = {
+    0x00, 0xff, 0xff, 0x00, 0x00, 0xff, 0xff, 0x00, 0xff, 0xff, 0xff, 0xff,
+    0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff,
+    0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x00, 0x00, 0x00, 0xff, 0xff,
+    0xff, 0xff, 0x00, 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0x00, 0x00, 0x00
+};
+
+// 10x8
+static const uint8_t PROGMEM heartBig[] = {
+    0x00, 0x8a, 0xff, 0x71, 0x17, 0x0d, 0x4c, 0xff, 0xc1, 0x05, 0x7c, 0xe0,
+    0xff, 0xd3, 0x90, 0x8a, 0xc0, 0xff, 0xf3, 0x90, 0xff, 0xff, 0xff, 0xff,
+    0xf3, 0xf1, 0xfd, 0xff, 0xfe, 0xff, 0xa1, 0xff, 0xff, 0xff, 0xff, 0xff,
+    0xff, 0xff, 0xff, 0xaf, 0x11, 0x8d, 0xf7, 0xff, 0xff, 0xff, 0xff, 0xff,
+    0xb8, 0x1b, 0x00, 0x0f, 0x87, 0xf3, 0xff, 0xff, 0xf8, 0xa8, 0x1c, 0x00,
+    0x00, 0x00, 0x13, 0x89, 0xf9, 0xfc, 0xa1, 0x24, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x22, 0xd1, 0xea, 0x3c, 0x00, 0x00, 0x00
+};
