--- src/main.h
@@ -1,1 +1,0 @@
-#include <Arduino.h>
